requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bitarray==3.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.1.8
contourpy==1.3.0
cryptography==44.0.3
cycler==0.12.1
cytoolz==1.1.0
datasets==3.1.0
dill==0.3.8
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
flwr==1.23.0
flwr-datasets==0.5.0
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2024.9.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-health-checking==1.62.3
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.3
idna==3.11
imbalanced-learn==0.12.4
importlib_resources==6.5.2
iterators==0.0.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pandas==2.3.3
parsimonious==0.10.0
pathspec==0.12.1
pillow==11.3.0
playwright==1.58.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
ray==2.31.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==13.9.4
rlp==4.1.0
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.4.0
tomli_w==1.2.0
toolz==1.1.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.12.5
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
web3==7.14.1
websockets==15.0.1
wfdb==4.3.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0